Move company bank account onboarding to settings
